Restructure code to make it working as a pip package. Bump the version to 1.1.1 to be same across github and pip. Update README to include instructions on how to use the package from pip
