+++ README.md
@@ -1,12 +1,22 @@
 # TODO Notifier
 
+- [TODO Notifier](#todo-notifier)
+  - [Description](#description)
+  - [How to use?](#how-to-use)
+    - [Method 1: As a pip package](#method-1-as-a-pip-package)
+    - [Method 2: Directly cloning from GitHub](#method-2-directly-cloning-from-github)
+  - [Technical Details](#technical-details)
+    - [Working](#working)
+    - [Salient Features](#salient-features)
+    - [Other Salient Features](#other-salient-features)
+
 ## Description
 
 More often than not, we put some TODO items in code and forget about them. Sometimes, we think of coming back to a TODO item by some date but miss it being too busy with some other development.
 
-TODO Notifier aims to solve this problem. It parses through any project, collects all the todo items and send automated reminders about them.
+TODO Notifier aims to solve this problem. It parses through any project, collects all the todo items and send automated reminders about them *[Upcoming feature]*.
 
-Recommended format to write TODO items
+Recommended format to write TODO items**
 
 `TODO {2022-05-22} @user_name msg`
 
@@ -21,25 +31,44 @@ However, the relative position of the `TODO` item, date inside `{}` brackets, us
 
 The framework is robust in the sense that if the TODO item misses some data like date and/or message and/or username etc., the respective TODO item will still be picked up by the TODO Notifier. However without relevant information, certain functionalities may not work. For e.g. without date, it cannot know if the TODO item has overshoot its expected date of completion.
 
-Example to dry run the code on a local repository:
+## How to use?
+
+### Method 1: As a pip package
+
+Install TODO Notifier by running `pip install todonotifier`
+
+Then you can use below code to try the TODO Notifier.
 
 ```python
-url = ""  # Full local address of the project
-project_dir_name = ""  # name of the project
-config = DefaultConfig(save_html_reports=True, ignore_todo_case=True)
-connect = Connect(connect_method=CONNECT_METHOD.DRY_RUN_DIR, project_dir_name=project_dir_name, url=url)
+from todonotifier.config import DefaultConfig
+from todonotifier.connect import CONNECT_METHOD, Connect
+from todonotifier.driver import run as driver_run
 
-driver_run(connect=connect, config=config)
+git_url: str = ""  # Placeholder for HTTPS/SSH based git url. It could also be a local folder address
+project_dir_name: str = ""  # Placeholder. Must match project name in git repository/local folder
+
+config = DefaultConfig(save_html_reports=True, ignore_todo_case=True)  # Change per need
+connect = Connect(connect_method=CONNECT_METHOD.GIT_CLONE, project_dir_name=project_dir_name, url=git_url, branch_name="production")  # If using a local folder address in `git_url`, then change `connect_method` to `CONNECT_METHOD.DRY_RUN_DIR`
+
+driver_run(connect=connect, config=config)  # This would generate (by default) 3 summary files in `.report` folder in current working directory
 ```
 
-## Technical details
+### Method 2: Directly cloning from GitHub
+
+Clone using `git clone https://github.com/ashu-tosh-kumar/todo_notifier.git`
+
+Then you can use the `user_driver.py` file to run it.
+
+Also, if you plan to use it simply by cloning the repository, you can put your custom code in `user_driver.py` as the same will not be updated and is added for end users.
+
+## Technical Details
 
 ### Working
 
 - TODO Notifier copies/clones the respective repository into a temporary location to avoid the risk of modifying any file.
 - It then reads through all the files in the project and collects all the TODO items
 - It then generates the summaries as specified in the configuration
-- Finally it sends the notifications (by default via Emails) to respective users and a group as a whole [**NOT YET IMPLEMENTED**].
+- Finally it sends the notifications (by default via Emails) to respective users and a group as a whole *[Upcoming feature]*.
 
 ### Salient Features
