+++ todonotifier/constants.py
